refactor/809 - Replace old WaypointModel with new
