+++ src/assets/scripts/client/aircraft/FlightManagementSystem/WaypointModel.js
@@ -1,4 +1,6 @@
-import _get from 'lodash/get';
+import _isArray from 'lodash/isArray';
+import _isEmpty from 'lodash/isEmpty';
+import FixCollection from '../../navigationLibrary/Fix/FixCollection';
 import {
     INVALID_INDEX,
     INVALID_NUMBER
@@ -7,206 +9,49 @@ import {
     RNAV_WAYPOINT_DISPLAY_NAME,
     RNAV_WAYPOINT_PREFIX
 } from '../../constants/waypointConstants';
-import { extractHeadingFromVectorSegment } from '../../navigationLibrary/Route/routeStringFormatHelper';
-import { degreesToRadians } from '../../utilities/unitConverters';
+// import { extractHeadingFromVectorSegment } from '../../navigationLibrary/Route/routeStringFormatHelper';
+import {
+    degreesToRadians,
+    DECIMAL_RADIX
+} from '../../utilities/unitConverters';
 
 /**
- * A representation of navigation point within a flight plan.
+ * A navigation point within an aircraft's flight plan
  *
- * // TODO: needs more info here
- * This navigation point can originate from one of several sources:
- * - `FixModel`, when flying to a specific fix or holding at a specific fix
- * - `StandardRouteWaypointModel`, when flying a standardRoute (SID/STAR)
+ * This may include various types of restrictions or holding information, all
+ * of which are used by the aircraft to follow various routes and procedures
+ * utilized by the controller.
  *
  * @class WaypointModel
  */
 export default class WaypointModel {
     /**
-     *
-     * @constructor
      * @for WaypointModel
-     * @param waypointProps {object}
+     * @constructor
      */
-    constructor(waypointProps) {
-        /**
-        * Maximum altitude at which to cross this waypoint
-        *
-        * @for WaypointModel
-        * @property altitudeMaximum
-        * @type {number}
-        * @default -1
-        */
-        this.altitudeMaximum = INVALID_NUMBER;
-
-        /**
-        * Minimum altitude at which to cross this waypoint
-        *
-        * @for WaypointModel
-        * @property altitudeMinimum
-        * @type {number}
-        * @default INVALID_NUMBER
-        */
-        this.altitudeMinimum = INVALID_NUMBER;
-
-        /**
-        * Flag used to determine if a waypoint is for a holding pattern
-        *
-        * Typically used from the fms as `fms#currentWaypoint`
-        *
-        * @for WaypointModel
-        * @property isHold
-        * @type {boolean}
-        * @default false
-        */
-        this.isHold = false;
-
-        /**
-         * Maximum speed at which to cross this waypoint
-         *
-         * @for WaypointModel
-         * @property speedMaximum
-         * @type {number}
-         * @default INVALID_NUMBER
-         */
-        this.speedMaximum = INVALID_NUMBER;
-
-        /**
-         * Minimum speed at which to cross this waypoint
-         *
-         * @for WaypointModel
-         * @property speedMinimum
-         * @type {number}
-         * @default INVALID_NUMBER
-         */
-        this.speedMinimum = INVALID_NUMBER;
-
-        // TODO: Should this just be `INVALID_NUMBER`?
-        /**
-         * Timer id for holding pattern
-         *
-         * Used only when waypoint is a holding pattern
-         *
-         * @for WaypointModel
-         * @property timer
-         * @type {number}
-         * @default -999
-         * @private
-         */
-        this.timer = -999;
-
-        /**
-        * Heading to fly during the inbound leg of a holding pattern at this fix
-        *
-        * @for WaypointModel
-        * @property _holdingPatternInboundHeading
-        * @type {number}
-        * @default INVALID_NUMBER
-        * @private
-        */
-        this._holdingPatternInboundHeading = INVALID_NUMBER;
-
-        /**
-         * Flag used to determine if the waypoint must be flown over before the
-         * aircraft may proceed to the next fix on their route.
-         *
-         * @for WaypointModel
-         * @property _isFlyOverWaypoint
-         * @type {boolean}
-         * @default false
-         * @private
-         */
-        this._isFlyOverWaypoint = false;
-
-        /**
-         * Flag used to determine if a waypoint is for a vector
-         *
-         * @for WaypointModel
-         * @property _isVector
-         * @type {boolean}
-         * @default false
-         * @private
-         */
-        this._isVector = false;
-
-        /**
-         * Length of each leg in holding pattern.
-         *
-         * Measured in either minutes or nautical miles
-         * Used only when waypoint is a holding pattern
-         *
-         * @for WaypointModel
-         * @property _legLength
-         * @type {string}
-         * @default ''
-         * @private
-         */
-        this._legLength = '';
-
-        /**
-        * Name of the waypoint
-        *
-        * Should be an ICAO identifier
-        *
-        * @for WaypointModel
-        * @property name
-        * @type {string}
-        * @default ''
-        * @private
-        */
-        this._name = '';
-
-        /**
-        * `StaticPositionModel` of the waypoint.
-        *
-        * @for WaypointModel
-        * @property _positionModel
-        * @type {StaticPositionModel}
-        * @default null
-        * @private
-        */
-        this._positionModel = null;
-
-        /**
-         * Direction to turn for a holding pattern
-         *
-         * Used only when waypoint is a holding pattern
-         *
-         * @for WaypointModel
-         * @property _turnDirection
-         * @type {string}
-         * @default ''
-         * @private
-         */
-        this._turnDirection = '';
-
-        this.init(waypointProps);
-    }
-
-    /**
-     * Returns the name of the waypoint
-     *
-     * Will return `RNAV` if the waypoint is a specific point in space
-     * and not a named fixed. These waypoints are prefixed with a
-     * `_` symbol.
-     *
-     * @property name
-     * @type {string}
-     * @return {string}
-     */
-    get name() {
-        if (this._name.indexOf(RNAV_WAYPOINT_PREFIX) !== INVALID_INDEX) {
-            return RNAV_WAYPOINT_DISPLAY_NAME;
+    constructor(data) {
+        if (typeof data !== 'string' && !_isArray(data)) {
+            throw new TypeError(`Expected valid data to create WaypointModel but received ${data}`);
         }
 
-        return this._name;
-    }
+        this.altitudeMaximum = -1;
+        this.altitudeMinimum = -1;
+        this.speedMaximum = -1;
+        this.speedMinimum = -1;
+        /**
+         * Contains `timer`, `inboundHeading`, 'legLength', 'turnDirection'
+         * @for WaypointModel
+         * @property _holdParameters
+         * @type {object}
+         */
+        this._holdParameters = null;
+        this._isFlyOverWaypoint = false;
+        this._isHoldWaypoint = false;
+        this._isVectorWaypoint = false;
+        this._name = '';
+        this._positionModel = null;
 
-    /**
-     * @property name
-     * @type {string}
-     */
-    set name(nameUpdate) {
-        this._name = name;
+        this.init(data);
     }
 
     /**
@@ -246,15 +91,63 @@ export default class WaypointModel {
      * @property hold
      * @return {object}
      */
-    get hold() {
-        return {
-            dirTurns: this._turnDirection,
-            fixName: this._name,
-            fixPos: this._positionModel.relativePosition,
-            inboundHeading: this._holdingPatternInboundHeading,
-            legLength: parseInt(this._legLength.replace('min', ''), 10),
-            timer: this.timer
-        };
+    get holdParameters() {
+        if (!this._isHoldWaypoint) {
+            return;
+        }
+
+        return this._holdParameters;
+    }
+
+    /**
+     * Returns whether `this` is a fly-over waypoint
+     * @for WaypointModel
+     * @property isFlyOverWaypoint
+     * @return {boolean}
+     */
+    get isFlyOverWaypoint() {
+        return this._isFlyOverWaypoint;
+    }
+
+    /**
+    * Returns whether `this` is a hold waypoint
+    *
+    * @for WaypointModel
+    * @property isHoldWaypoint
+    * @type {boolean}
+    */
+    get isHoldWaypoint() {
+        return this._isHoldWaypoint;
+    }
+
+    /**
+    * Returns whether `this` is a vector waypoint
+    *
+    * @for WaypointModel
+    * @property isVector
+    * @return {boolean}
+    */
+    get isVectorWaypoint() {
+        return this._isVectorWaypoint;
+    }
+
+    /**
+     * Returns the name of the waypoint
+     *
+     * Will return `RNAV` if the waypoint is a specific point in space
+     * and not a named fixed. These waypoints are prefixed with a
+     * `_` symbol.
+     *
+     * @property name
+     * @type {string}
+     * @return {string}
+     */
+    get name() {
+        if (this._name.indexOf(RNAV_WAYPOINT_PREFIX) !== INVALID_INDEX) {
+            return RNAV_WAYPOINT_DISPLAY_NAME;
+        }
+
+        return this._name;
     }
 
     /**
@@ -279,92 +172,146 @@ export default class WaypointModel {
         return this._positionModel.relativePosition;
     }
 
+    // ------------------------------ LIFECYCLE ------------------------------
+
     /**
-     * Returns whether `this` is a fly-over waypoint
+     * Initialize class properties
+     *
      * @for WaypointModel
-     * @property isFlyOverWaypoint
-     * @return {boolean}
+     * @method init
+     * @param data {object}
+     * @chainable
      */
-    get isFlyOverWaypoint() {
-        return this._isFlyOverWaypoint;
+    init(data) {
+        let fixName = data;
+        let restrictions = '';
+
+        if (_isArray(data)) {
+            if (data.length !== 2) {
+                throw new TypeError(`Expected restricted fix to have restrictions, but received ${data}`);
+            }
+
+            fixName = data[0];
+            restrictions = data[1];
+        }
+
+        this._name = fixName.replace('@', '').replace('^', '');
+
+        this._initSpecialWaypoint(fixName);
+        this._applyRestrictions(restrictions);
+        this._initializePosition();
+
+        return;
     }
 
     /**
-     * Returns whether `this` is a vector waypoint
+     * Reset class properties
      *
      * @for WaypointModel
-     * @property isVector
-     * @return {boolean}
+     * @method reset
+     * @chainable
      */
-    get isVector() {
-        return this._isVector;
+    reset() {
+        this.altitudeMaximum = -1;
+        this.altitudeMinimum = -1;
+        this.speedMaximum = -1;
+        this.speedMinimum = -1;
+        this._holdParameters = null;
+        this._isFlyOverWaypoint = false;
+        this._isHoldWaypoint = false;
+        this._isVectorWaypoint = false;
+        this._name = '';
+        this._positionModel = null;
+
+        return this;
     }
+
+    /**
+     * Initialize properties to make this waypoint a fly-over waypoint
+     *
+     * @for WaypointModel
+     * @method _initFlyOverWaypoint
+     * @private
+     */
+    _initFlyOverWaypoint() {
+        this._isFlyOverWaypoint = true;
+    }
+
+    /**
+     * Initialize properties to make this waypoint a hold waypoint
+     *
+     * @for WaypointModel
+     * @method _initHoldWaypoint
+     * @private
+     */
+    _initHoldWaypoint() {
+        this._isHoldWaypoint = true;
+        // FIXME: These should be coming from a const file somewhere instead of being hard-coded
+        this._holdParameters = {
+            inboundHeading: undefined,
+            legLength: 1,
+            timer: INVALID_NUMBER,
+            turnDirection: 'right'
+        };
+    }
+
+    /**
+     * Perform additional initialization tasks as needed if waypoint is a flyover/hold/vector/etc waypoint
+     *
+     * @for WaypointModel
+     * @method _initSpecialWaypoint
+     * @param fixname {string} name of the fix, including any special characters
+     */
+    _initSpecialWaypoint(fixName) {
+        if (fixName.indexOf('^') !== INVALID_INDEX) {
+            this._initFlyOverWaypoint();
+
+            return;
+        }
+
+        if (fixName.indexOf('@') !== INVALID_INDEX) {
+            this._initHoldWaypoint();
+
+            return;
+        }
+
+        if (fixName.indexOf('#') !== INVALID_INDEX) {
+            this._initVectorWaypoint();
+
+            return;
+        }
+    }
+
+    /**
+     * Initialize properties to make this waypoint a vector waypoint
+     *
+     * @for WaypointModel
+     * @method _initVectorWaypoint
+     * @private
+     */
+    _initVectorWaypoint() {
+        this._isVectorWaypoint = true;
+    }
+
+    // ------------------------------ PUBLIC ------------------------------
 
     /**
      * When `#_isVector` is true, this gets the heading that should be flown
      *
      * @for WaypointModel
-     * @property vector
+     * @method _getVector
      * @type {number}
      */
-    get vector() {
-        if (!this.isVector) {
+    getVector() {
+        if (!this._isVectorWaypoint) {
             return;
         }
 
-        const headingInDegrees = parseInt(extractHeadingFromVectorSegment(this._name), 10);
+        const fixNameWithOutPoundSign = this._name.replace('#', '');
+        const headingInDegrees = parseInt(fixNameWithOutPoundSign, DECIMAL_RADIX);
         const headingInRadians = degreesToRadians(headingInDegrees);
 
         return headingInRadians;
-    }
-
-    /**
-     * Initialize the class properties
-     *
-     * Should be run only on instantiation
-     *
-     * @For WaypointModel
-     * @method init
-     * @param waypointProps {object}
-     */
-    init(waypointProps) {
-        this._name = waypointProps.name.toLowerCase();
-        this._positionModel = waypointProps.positionModel;
-        this.speedMaximum = parseInt(waypointProps.speedMaximum, 10);
-        this.speedMinimum = parseInt(waypointProps.speedMinimum, 10);
-        this.altitudeMaximum = parseInt(waypointProps.altitudeMaximum, 10);
-        this.altitudeMinimum = parseInt(waypointProps.altitudeMinimum, 10);
-        this._isFlyOverWaypoint = waypointProps.isFlyOverWaypoint;
-        this._isVector = waypointProps.isVector;
-
-        // these properties will only be available for holding pattern waypoints
-        this.isHold = _get(waypointProps, 'isHold', this.isHold);
-        this.timer = _get(waypointProps, 'timer', this.timer);
-        this._holdingPatternInboundHeading = _get(waypointProps, '_holdingPatternInboundHeading',
-            this._holdingPatternInboundHeading
-        );
-        this._legLength = _get(waypointProps, 'legLength', this._legLength);
-        this._turnDirection = _get(waypointProps, 'turnDirection', this._turnDirection);
-    }
-
-    /**
-     * Tear down the instance and reset class properties
-     *
-     * @for WaypointModel
-     * @method destroy
-     */
-    destroy() {
-        this._name = '';
-        this._turnDirection = '';
-        this._legLength = '';
-        this._positionModel = null;
-
-        this.isHold = false;
-        this.speedMaximum = INVALID_NUMBER;
-        this.speedMinimum = INVALID_NUMBER;
-        this.altitudeMaximum = INVALID_NUMBER;
-        this.altitudeMinimum = INVALID_NUMBER;
-        this.timer = -999;
     }
 
     /**
@@ -394,18 +341,120 @@ export default class WaypointModel {
     }
 
     /**
-     * Add hold-specific properties to an existing `WaypointModel` instance
+     * Stores provided parameters for holding pattern, and marks this as a hold waypoint
      *
      * @for WaypointModel
-     * @method updateWaypointWithHoldProps
-     * @param inboundHeading {number}  in radians
-     * @param turnDirection {string}   either left or right
-     * @param legLength {string}       length of the hold leg in minutes or nm
+     * @method setHoldParametersAndArmHold
+     * @param inboundHeading {number} in radians
+     * @param turnDirection {string} either left or right
+     * @param legLength {string} length of the hold leg in minutes or nm
      */
-    updateWaypointWithHoldProps(inboundHeading, turnDirection, legLength) {
-        this.isHold = true;
-        this._holdingPatternInboundHeading = inboundHeading;
-        this._turnDirection = turnDirection;
-        this._legLength = legLength;
+    setHoldParametersAndArmHold(inboundHeading, turnDirection, legLength) {
+        this._isHoldWaypoint = true;
+        this._holdParameters = {
+            turnDirection: turnDirection,
+            inboundHeading: inboundHeading,
+            legLength: legLength,
+            timer: INVALID_NUMBER
+        };
+    }
+
+    // ------------------------------ PRIVATE ------------------------------
+
+    /**
+     * Apply an altitude restriction in the appropriate properties
+     *
+     * @for WaypointModel
+     * @method _applyAltitudeRestriction
+     * @param restriction {string}
+     */
+    _applyAltitudeRestriction(restriction) {
+        const altitude = parseInt(restriction, 10) * 100;
+
+        if (restriction.indexOf('+') !== -1) {
+            this.altitudeMinimum = altitude;
+
+            return;
+        } else if (restriction.indexOf('-') !== -1) {
+            this.altitudeMaximum = altitude;
+
+            return;
+        }
+
+        this.altitudeMaximum = altitude;
+        this.altitudeMinimum = altitude;
+    }
+
+    /**
+     * Parse the restrictions, and store the inferred meaning in the appropriate properties
+     *
+     * @for WaypointModel
+     * @method _applyRestrictions
+     * @param restrictions {string} restrictions, separated by pipe symbol: '|'
+     */
+    _applyRestrictions(restrictions) {
+        if (_isEmpty(restrictions)) {
+            return;
+        }
+
+        const restrictionCollection = restrictions.split('|');
+
+        for (let i = 0; i < restrictionCollection.length; i++) {
+            const restriction = restrictionCollection[i];
+
+            // looking at the first letter of a restriction
+            if (restriction[0] === 'A') {
+                this._applyAltitudeRestriction(restriction.substr(1));
+            } else if (restriction[0] === 'S') {
+                this._applySpeedRestriction(restriction.substr(1));
+            } else {
+                throw new TypeError('Expected "A" or "S" prefix on restriction, ' +
+                    `but received prefix '${restriction[0]}'`);
+            }
+        }
+    }
+
+    /**
+     * Apply a speed restriction in the appropriate properties
+     *
+     * @for WaypointModel
+     * @method _applySpeedRestriction
+     * @param restriction {string}
+     */
+    _applySpeedRestriction(restriction) {
+        const speed = parseInt(restriction, 10);
+
+        if (restriction.indexOf('+') !== -1) {
+            this.speedMinimum = speed;
+
+            return;
+        } else if (restriction.indexOf('-') !== -1) {
+            this.speedMaximum = speed;
+
+            return;
+        }
+
+        this.speedMaximum = speed;
+        this.speedMinimum = speed;
+    }
+
+    /**
+     * Initialize the waypoint's position model based on #_name
+     *
+     * @for WaypointModel
+     * @method _initializePosition
+     */
+    _initializePosition() {
+        if (this._isVectorWaypoint) {
+            return;
+        }
+
+        const fixPosition = FixCollection.getPositionModelForFixName(this._name);
+
+        if (fixPosition === null) {
+            throw new TypeError(`Expected fix with known position, but cannot find fix '${this._name}'`);
+        }
+
+        this._positionModel = fixPosition;
     }
 }
