+++ test/aircraft/FlightManagementSystem/LegModel.spec.js
@@ -4,7 +4,7 @@ import sinon from 'sinon';
 // import _isEqual from 'lodash/isEqual';
 import LegModel from '../../../src/assets/scripts/client/aircraft/FlightManagementSystem/LegModel';
 import ProcedureDefinitionModel from '../../../src/assets/scripts/client/navigationLibrary/Procedure/ProcedureDefinitionModel';
-import ProcedureWaypointModel from '../../../src/assets/scripts/client/navigationLibrary/Procedure/ProcedureWaypointModel';
+// import WaypointModel from '../../../src/assets/scripts/client/aircraft/FlightManagementSystem/WaypointModel';
 import NavigationLibrary from '../../../src/assets/scripts/client/navigationLibrary/NavigationLibrary';
 import { AIRPORT_JSON_KLAS_MOCK } from '../../airport/_mocks/airportJsonMock';
 // import { HOLD_AT_PRESENT_LOCATION_MOCK } from '../_mocks/aircraftMocks';
